backend/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
streamlit==1.31.0

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 数据处理
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# 文件解析（PDF/Word）
pymupdf==1.23.8          # PDF解析（PyMuPDF）
pdfplumber==0.10.0       # PDF解析备选
python-docx==0.8.11      # Word文档解析

# AI/ML
scikit-learn==1.4.0
xgboost==2.0.3
shap==0.43.0
joblib==1.3.2

# LLM 调用
httpx==0.26.0
openai==1.10.0

# 支付
alipay-sdk==3.3.1

# 可视化
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# 工具
jinja2==3.1.3
pydantic==2.5.3
pydantic-settings==2.1.0
cachetools==5.3.2
extra-streamlit-components==0.1.56
